// src/app/modelo/reserva-dto.ts
export class ReservaDTO {

    id:string = "";
    cliente: string = "";
    regimen: number = 0;
    fechaInicio: string = "";
    fechaSalida: string = "";
    estado: string = "";
    fechaEntrega: string = "";
    precio: number = 0;
    idHotel : string = "";
    paqueteHotel: string = "";
    paqueteTour: string = "";
    idAutomovil: string = "";

}